style: cat user_add.go user_delete.go user_edit.go user_list.go > user.go
